requirements.txt
asgiref==3.4.1
Brotli==1.0.9
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.5
click==8.0.1
colorama==0.4.4
ConfigArgParse==1.5.2
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2.7
django-allauth==0.45.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
Flask==2.0.1
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==21.8.0
geventhttpclient==1.5.1
greenlet==1.1.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
locust==2.2.2
MarkupSafe==2.0.1
msgpack==1.0.2
oauthlib==3.1.1
psutil==5.8.0
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
pywin32==301
pyzmq==22.2.1
requests==2.26.0
requests-oauthlib==1.3.0
roundrobin==0.0.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.6
Werkzeug==2.0.1
zope.event==4.5.0
zope.interface==5.4.0
